+++ crates/libneo-gpui/src/platform/mac/mod.rs
@@ -6,6 +6,7 @@ mod handle;
 pub(crate) mod menu;
 pub(crate) mod table;
 pub(crate) mod visual_effect;
+pub(crate) mod window;
 
 use objc2::MainThreadMarker;
 use objc2_foundation::{NSOperatingSystemVersion, NSProcessInfo};